move code form dunder init to models
